backend/requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
defusedxml==0.7.1
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.135.1
firebase_admin==7.2.0
fonttools==4.62.0
fpdf2==2.8.7
google-api-core==2.30.0
google-auth==2.49.0
google-cloud-aiplatform==1.140.0
google-cloud-bigquery==3.40.1
google-cloud-core==2.5.0
google-cloud-firestore==2.24.0
google-cloud-pubsub==2.35.0
google-cloud-resource-manager==1.16.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-genai==1.66.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
grpc-google-iam-v1==0.14.3
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
msgpack==1.1.2
opentelemetry-api==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
packaging==26.0
pillow==12.1.1
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
zipp==3.23.0
